requirements.txt
newspaper3k==0.2.8
requests==2.23.0
